Promote ash to first-class src/agent/
ash is a load-bearing built-in extension (peer to shell), not one of the
peripheral built-ins. Move the activate function from
extensions/agent-backend.ts into src/agent/index.ts alongside the
existing module re-exports, and relocate the provider built-ins (used
only by ash) to src/agent/providers/. src/extensions/ now contains only
genuinely cross-cutting built-ins.
