+++ src/agent/providers/openai-compatible.ts
