+++ src/agent/providers/openai.ts
