+++ src/extensions/index.ts
@@ -22,18 +22,18 @@ export const BUILTIN_EXTENSIONS: Array<{
   load: () => Promise<ActivateFn>;
 }> = [
   { name: "shell-context",   load: () => import("../shell/shell-context.js").then(m => m.default) },
-  { name: "agent-backend",    load: () => import("./agent-backend.js").then(m => m.default) },
+  { name: "agent-backend",    load: () => import("../agent/index.js").then(m => m.default) },
   { name: "openrouter",
     when: () => !!resolveApiKey("openrouter").key,
-    load: () => import("./providers/openrouter.js").then(m => m.default) },
+    load: () => import("../agent/providers/openrouter.js").then(m => m.default) },
   { name: "openai",
     when: () => !!resolveApiKey("openai").key && !process.env.OPENAI_BASE_URL,
-    load: () => import("./providers/openai.js").then(m => m.default) },
+    load: () => import("../agent/providers/openai.js").then(m => m.default) },
   { name: "openai-compatible",
     when: () => !!process.env.OPENAI_BASE_URL,
-    load: () => import("./providers/openai-compatible.js").then(m => m.default) },
+    load: () => import("../agent/providers/openai-compatible.js").then(m => m.default) },
   { name: "deepseek",
-    load: () => import("./providers/deepseek.js").then(m => m.default) },
+    load: () => import("../agent/providers/deepseek.js").then(m => m.default) },
   { name: "tui-renderer",     load: () => import("../shell/tui-renderer.js").then(m => m.default) },
   { name: "slash-commands",    load: () => import("./slash-commands.js").then(m => m.default) },
   { name: "file-autocomplete", load: () => import("../shell/file-autocomplete.js").then(m => m.default) },
